+++ RELEASE_NOTES.md
@@ -7,6 +7,7 @@
 - Added "Invalid" status for competitions, e.g. if a channel is on longer valid.
 - Added validation section for competitions and new status type "Invalid". This way, errors can be displayed, e.g. if a competition was created for a game or Discord channel that no longer exists.
 - Added "VPS Link" and "Download Link" fields for online competitions. If set, this information will be rendered with the competition card and metadata, so players can download the correct version of the table.
+- Added "Channel" column to overview table to have a better overview when multiple competitions are on one server.
 
 **Preferences: Discord Bot**